Vector plot
Inside utils.py I added a small code to plot a vector field. Right now,
it computes the normal vector N and plots some vectors at certain points
in the imshow image. The code to show the plots is "right", but the
computation of the vector directions may not be so. Looking at the
radial field, the results don't add up, but I feel it's mostly due to
the fact that I'm not using the most up to date results.